Added LineId at boarding/alighting
Added LineIdAtBoarding in LegBoardStructure and LineIdAtAlighting in LegAlightStructure. This way, an arbitrary identifier for a service at boarding and alighting can be transported, which is independent from JourneyRef. This can be used in distributed environments in order to recognize, that two systems refer the same line/service while they still use their own internal references.