Added ProtoProduct into FareProductStructure (#60)
A proto product is a fare product, which shall not be shown to the user directly. Instead it carries fare product related information, which can be processed further. Can be used for tariffing in a distributed environment (e.g. EU-Spirit), where OJP services can deliver only parts of a fare product.