Added Interchange IDs at boarding/alighting (#61)
Added DistributorInterchangeId in LegBoardStructure and FeederInterchangeId in LegAlightStructure. This way, an arbitrary identifier for feeding service at alighting and distributing service at boarding can be transported, which is independent from JourneyRef. Using the same interchange IDs means that the interchange is in fact a stay-seated scenario. This can be used in distributed environments (e.g. EU-Spirit) in order to recognize, that two systems refer the same line/service while they still use their own internal references.