Added ReferredSystemId in TopographicPlaceStructure (#62)
In an distributed environment (e.g. EU-Spirit), the process of place identification can be a two-steps process. In the first step, a topographic place (e.g. city, municipality) is identified from the user's input. In the second step, the system related to the topographic place is queried for places. In order to do so, the topographic places from the first step need to carry the information, to which system they relate.