Integrated changes from pvgrumbokw
Please change three things:
a) The switches have effect on the content included in the response. Therefore they should be part of the TripContentFilterGroup.
b) The switches in the TripContentFilterGroup are always include switches. So please define them as IncludeXXX with the default set to true.
c) Please consider also the other OJP services. Where appropriate these switches should also be included into the content filter groups of those requests.