Renaming of Location with Place (#82)
According to TRANSMODEL a location is only a geographical position, while a place has a location along with other attributes. Within OJP the two terms were used inconsistently. This commit distinguishes the two terms properly. In most cases the term location was replaced by the term place. This applies to type and element names but also to annotations. This commit addresses issue #44.