Changes in published name for services in ServiceGroup (#89)
Previously the published name of a service should have been stored in the element "PublishedLineName" in  ServiceGroup. As the term line is misleading here, the element is renamed to "PublishedServiceName". It shall contain the name of a service which is communicated to the customer.
Additionally a new optional element "TrainNumber" is introduced which can store the number of a service (usually a train) according to NeTEx.