Added content filter for stop hierarchy into StopEvent (#96)
Added a simple type HierarchyEnumeration. Added a content filter IncludeStopHierarchy into the StopEventContentFilterGroup. This way a requestor can choose, which parts of the stop hierarchy of the referenced place will be included into the StopEvent response. Default is that the server determines which related stops are included, which is the legacy behaviour.