Consistent use of WebLinkStructure
Removed 3 occurrences of xs:anyURI and exchanged it with WebLinkStructure (MultiTripFareResultStructure and BookingArrangementStructure). Introduced a WebLinkGroup consisting of URL, MIME type and embeddedness. Used that group within the WebLinkStructure (without actually changing it). Exchanged xs:anyURI in GeneralAttributeStructure by the WebLinkGroup as there is already a text available for the link.