Added service arrival and departure structures (#156)
Introduced within OJP_JourneySupport.xsd the complex types ServiceArrivalStructure and ServiceDepartureStructure. Changed within CallAtStopStructure, LegBoardStructure, LegAlightStructure, and LegIntermediateStructure the respective element ServiceArrival to the type ServiceArrivalStructure and the element ServiceDeparture to the type ServiceDepartureStructure. Main reason was to eliminate structured elements within an appropriate definition of a complex type, in order to enable automatic documentation generation.