adapt to changes discussed in the group.
This will however, change the intension Andreas had here. But I guess this is unavoidable as there are differences in opinion. And VehicleNumber has a different meaning and also does not exist in the way Andreas intended in Transmodel. Using TRAIN NUMBER also for non-train modes, if the service level intention is the some as fulfilled by TRAIN NUMBER seems better.