Example update to have a proper use of namespace
Following the proposal of namespace updated in OJP, as pointed by Matthias, this has consequences in all example.
Having a look arround them, the use of namespace is a bit of a mess in OJP.
Basically, headers should move to something like this in examples:
<OJP xmlns="http://www.vdv.de/ojp" xmlns:ojp="http://www.vdv.de/ojp" xmlns:siri="http://www.siri.org.uk/siri" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" version="1.1-dev" xsi:schemaLocation="http://www.vdv.de/ojp ../../OJP.xsd">

But you can see that there is 2 different uses of "http://www.vdv.de/ojp" because the examples don't make a consistent use of the "ojp" prefix. Also, once "ojp" becomes the default namespace, then you need to use the "siri" one when you use somethnig from SIRI (that was kind of reverse up to now because SIRI wans made default.

By the way, that's one of the good side of tools like XML Spy that prevent that kind of mistake when you use it yo create examples ;-)